backend/scripts/requirements_stage3.txt
# Python dependencies for Stage 3 - Visual Recognition Training Pipeline
# Install with: pip install -r requirements_stage3.txt

# Image processing
Pillow>=10.0.0
imagehash>=4.3.1

# Azure Custom Vision SDK
azure-cognitiveservices-vision-customvision>=3.1.0

# Web scraping (optional - for enhanced scraping)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
